work on workouts

+++ src/server/src/main/java/com/project/domain/valueobjects/Weight.java
@@ -18,7 +18,7 @@ public class Weight {
 
     protected Weight() {}
 
-    public static Weight from(int weight) {
+    public static Weight from(double weight) {
         return new Weight(weight);
     }
